implement https://github.com/saintlab/mobileapp_android/issues/358

+++ Omnom.Android/libraries/Omnom.Menu/src/main/java/com/omnom/android/menu/model/Details.java
@@ -35,7 +35,7 @@ public abstract class Details implements Parcelable {
 	}
 
 	@Nullable
-	public abstract int weight();
+	public abstract double weight();
 
 	@Nullable
 	public abstract int cookingTime();
